Implement MVP: it loads
