+++ LICENSE
@@ -1,7 +1,3 @@
-The software provided in this example, both the code lines demonstrated in the
-[Jupyter Notebook](01-kibble-zurek.ipynb) and in additional helper files, is
-licensed under the Apache License, Version 2.0 (see below).
-
                                  Apache License
                            Version 2.0, January 2004
                         http://www.apache.org/licenses/
