+++ assets/__demo_variables.css
@@ -1,0 +1,48 @@
+/*
+Copyright 2024 D-Wave
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+/*
+Variable declarations for all CSS files.
+
+Dash reads all css files contained in `/assets/` so no imports are necessary.
+*/
+
+:root {
+    --blue-bright: #03B8FF;
+    --blue-light: #2A7DE1;
+    --blue-dark: #074C91;
+    --blue-darker: #2d4376;;
+    --blue-darkest: #202239;
+    --red-light: #F57677;
+    --red-dark: #AA3A3C;
+    --orange: #FF7006;
+    --teal-light: #06ECDC;
+    --teal: #17BEBB;
+    --teal-dark: #17BEBB;
+    --teal-darker: #008C82;
+    --grey-lighter: #F5F7FB;
+    --grey-light: #DBDBDB;
+    --grey: #AAAAAA;
+    --grey-medium: #737373;
+    --grey-dark: #222222;
+    --box-shadow: 0 0 1rem rgba(66, 82, 121, 0.2);
+    --font: "proxima-nova", "Helvetica Neue", sans-serif;
+    --banner-height: 6.5rem;
+    --left-col-width: 29rem;
+    --problem-details-height: 8rem;
+    --title-section-height: 13.5rem;
+    /*** Add new variables here ***/
+}
