+++ assets/_tabs.css
@@ -1,0 +1,78 @@
+/*
+Copyright 2024 D-Wave
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+/* Style rules for tabs like the input and results tabs */
+
+nav {
+    height: 100%;
+}
+
+.mantine-Tabs-root {
+    height: 100%;
+}
+
+.mantine-Tabs-panel {
+    height: calc(100% - var(--banner-height));
+}
+
+.mantine-Tabs-list {
+    --tabs-list-gap: 1.5rem;
+    height: 100%;
+}
+
+.mantine-Tabs-list::before {
+    border: none;
+}
+
+.mantine-Tabs-tab {
+    font-size: 1.5rem;
+    font-weight: 400;
+    border-radius: 0;
+    border-bottom-width: 0;
+    border-top-width: 5px;
+    height: 100%;
+}
+
+button.mantine-Tabs-tab:disabled {
+    filter: unset;
+    opacity: 1;
+    color: #7E7F86;
+    border-color: transparent;
+}
+
+.mantine-Tabs-tab span {
+    color: #A0A1A5;
+    opacity: 1;
+    transition: color 0.2s ease-in-out;
+}
+
+.mantine-Tabs-tab:where([data-active]) span {
+    opacity: 1;
+    color: white;
+}
+
+.mantine-Tabs-tab:hover {
+    background-color: unset;
+}
+
+.mantine-Tabs-tab:hover:where(:not([data-active])) {
+    border-color: unset;
+}
+
+.mantine-Tabs-tab:hover:where(:not(:disabled)) span {
+    opacity: 1;
+    color: white;
+}
