--- helpers/tooltips.py
@@ -1,28 +1,0 @@
-# Copyright 2024 D-Wave
-#
-#    Licensed under the Apache License, Version 2.0 (the "License");
-#    you may not use this file except in compliance with the License.
-#    You may obtain a copy of the License at
-#
-#        http://www.apache.org/licenses/LICENSE-2.0
-#
-#    Unless required by applicable law or agreed to in writing, software
-#    distributed under the License is distributed on an "AS IS" BASIS,
-#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#    See the License for the specific language governing permissions and
-#    limitations under the License.
-
-tool_tips_kz_nm = {
-    "coupling_strength": "Coupling strength between spins in the ferromagnetic ring.",
-    "qpu_selection": "Quantum computers available to your account/project token.",
-    "quench_schedule_filename": """The fast-anneal schedule for the selected quantum computer.
-        If none exists, one from a different quantum computer is used (expect inaccuracies).""",
-}
-
-tool_tips_kz = {
-    "coupling_strength": """Coupling strength between spins in the ring.
-        Range of -2 (ferromagnetic) to +1 (anti-ferromagnetic).""",
-    "qpu_selection": "Quantum computers available to your account/project token.",
-    "quench_schedule_filename": """The fast-anneal schedule for the selected quantum computer.
-        If none exists, one from a different quantum computer is used (expect inaccuracies).""",
-}
