+++ requirements.txt
@@ -1,5 +1,5 @@
-dash[diskcache]~=3.0
-dash-bootstrap-components==1.5.0
+dash[diskcache]~=4.0
+dash-mantine-components~=2.5
 dwave-ocean-sdk~=9.3
 pandas~=2.3
