+++ static/1D_adiabatic.jpg
