+++ static/1D_finite.jpg
