+++ static/1D_kinks.jpg
