+++ static/formula_kink_density_operator.png
