+++ static/formula_kink_operator.png
