+++ static/spin_states.png
