+++ tests/test_cb_progress_bar.py
@@ -14,16 +14,16 @@
 
 from contextvars import copy_context
 
+from demo_configs import JOB_BAR_DISPLAY
 import pytest
 from dash._callback_context import context_value
 from dash._utils import AttributeDict
 
-from app import set_progress_bar
-from helpers.layouts_components import job_bar_display
+from demo_callbacks import set_progress_bar
 
 parametrize_vals = [
-    (f"{status}", job_bar_display[status][0], job_bar_display[status][1])
-    for status in job_bar_display.keys()
+    (f"{status}", JOB_BAR_DISPLAY[status][0], JOB_BAR_DISPLAY[status][1])
+    for status in JOB_BAR_DISPLAY.keys()
 ]
 parametrize_vals.extend([tuple(["BREAK FUNCTION", "exception", "exception"])])
